refactor: Move lead worktree tests to separate daemon_tests.rs file
Per CLAUDE.md convention, tests should be in separate files rather than
inline #[cfg(test)] mod tests blocks. Moves the two new tests
(test_lead_worktree_path_helper and test_worktree_manager_creates_lead_in_correct_location)
to daemon_tests.rs and fixes the misleading doc comment about CLI integration testing.
